requirements.txt
imutils==0.5.4
numpy==2.2.3
opencv-python==4.11.0.86
PyYAML==6.0.2
